+++ src/Resources/ResourceManager/Implementation/Policy/NewAzurePolicyAssignment.cs
@@ -24,14 +24,17 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Entities.Policy;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Entities.Resources;
     using Microsoft.Azure.Commands.ResourceManager.Cmdlets.Extensions;
-    using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
 
     using Newtonsoft.Json.Linq;
     using Policy;
+    using System.Collections.Generic;
+    using Microsoft.WindowsAzure.Commands.Common.CustomAttributes;
+    using Microsoft.Azure.Commands.Common.Exceptions;
 
     /// <summary>
     /// Creates a policy assignment.
     /// </summary>
+    [CmdletOutputBreakingChange(typeof(PsPolicyAssignment), NewOutputProperties = new String[] { "Identity" })]
     [Cmdlet(VerbsCommon.New, ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "PolicyAssignment", DefaultParameterSetName = PolicyCmdletBase.DefaultParameterSet), OutputType(typeof(PsPolicyAssignment))]
     public class NewAzurePolicyAssignmentCmdlet : PolicyCmdletBase, IDynamicParameters
     {
@@ -119,11 +122,21 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         [ValidateNotNullOrEmpty]
         public PolicyAssignmentEnforcementMode? EnforcementMode { get; set; }
 
-        /// <summary>
-        /// Gets or sets a flag indicating whether a system assigned identity should be added to the policy assignment.
-        /// </summary>
+        [CmdletParameterBreakingChange("AssignIdentity", ChangeDescription = "Parameter AssignIdentity is deprecated and will be removed in future releases. Please use the 'IdentityType' parameter instead.")]
         [Parameter(Mandatory = false, HelpMessage = PolicyHelpStrings.PolicyAssignmentAssignIdentityHelp)]
         public SwitchParameter AssignIdentity { get; set; }
+
+        /// <summary>
+        /// Gets or sets the type of managed identity to be assigned to the policy assignment.
+        /// </summary>
+        [Parameter(Mandatory = false, HelpMessage = PolicyHelpStrings.PolicyAssignmentIdentityTypeHelp)]
+        public ManagedIdentityType? IdentityType { get; set; }
+
+        /// <summary>
+        /// Gets or sets the ID of the user assigned managed identity to be assigned to the policy assignment.
+        /// </summary>
+        [Parameter(Mandatory = false, HelpMessage = PolicyHelpStrings.PolicyAssignmentIdentityIdHelp)]
+        public string IdentityId { get; set; }
 
         /// <summary>
         /// Gets or sets the location of the policy assignment. Only required when assigning a resource identity to the assignment.
@@ -160,6 +173,8 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
                 throw new PSInvalidOperationException("The supplied PolicySetDefinition object is invalid.");
             }
 
+            CheckIfIdentityPresent();
+
             string resourceId = GetResourceId();
 
             var apiVersion = string.IsNullOrWhiteSpace(this.ApiVersion) ? Constants.PolicyAssignmentApiVersion : this.ApiVersion;
@@ -195,14 +210,65 @@ namespace Microsoft.Azure.Commands.ResourceManager.Cmdlets.Implementation
         }
 
         /// <summary>
+        /// Verifies if identity type and/or identity ID and/or location are present
+        /// </summary>
+        private bool CheckIfIdentityPresent()
+        {
+            if (this.AssignIdentity.IsPresent && this.IdentityType != null) 
+            {
+                throw new AzPSArgumentException("Cannot specify both IdentityType and AssignIdentity at the same time.", "IdentityType");
+            }
+
+            if (this.AssignIdentity.IsPresent && !string.IsNullOrEmpty(this.IdentityId))
+            {
+                throw new AzPSArgumentException("Cannot specify both AssignIdentity and IdentityId at the same time.", "IdentityId");
+            }
+
+            if (this.IdentityType != null & string.IsNullOrEmpty(this.Location) || this.AssignIdentity.IsPresent && string.IsNullOrEmpty(this.Location))
+            {
+                throw new AzPSArgumentException("Location needs to be specified if a managed identity is to be assigned to the policy assignment.", "Location");
+            }
+
+            if (this.IdentityType == ManagedIdentityType.UserAssigned && string.IsNullOrEmpty(this.IdentityId))
+            {
+                throw new AzPSArgumentException("A user assigned identity id needs to be specified if the identity type is 'UserAssigned'.", "IdentityId");
+            }
+
+            if (this.IdentityType == ManagedIdentityType.SystemAssigned && !string.IsNullOrEmpty(this.IdentityId))
+            {
+                throw new AzPSArgumentException("Cannot specify an identity ID if identity type is 'SystemAssigned'.", "IdentityId");
+            }
+
+            return true;
+        }
+
+        /// <summary>
         /// Constructs the resource
         /// </summary>
         private JToken GetResource()
         {
+            if (this.AssignIdentity != null && this.AssignIdentity.IsPresent)
+            {
+                this.IdentityType = ManagedIdentityType.SystemAssigned;
+            }
+
+            ResourceIdentity identityObject = this.IdentityType != null ?
+                ( this.IdentityType == ManagedIdentityType.UserAssigned ?
+                    new ResourceIdentity
+                    {
+                        Type = IdentityType.ToString(),
+                        UserAssignedIdentities = new Dictionary<string, UserAssignedIdentityResource>
+                        {
+                            { this.IdentityId, new UserAssignedIdentityResource {} }
+                        }
+                    } :
+                    new ResourceIdentity { Type = IdentityType.ToString() }
+                ) : null;
+
             var policyassignmentObject = new PolicyAssignment
             {
                 Name = this.Name,
-                Identity = this.AssignIdentity.IsPresent ? new ResourceIdentity { Type = ResourceIdentityType.SystemAssigned } : null,
+                Identity = identityObject,
                 Location = this.Location,
                 Properties = new PolicyAssignmentProperties
                 {