+++ src/Compute/Compute.Test/ScenarioTests/VirtualMachineTests.cs
@@ -452,5 +452,12 @@ namespace Microsoft.Azure.Commands.Compute.Test.ScenarioTests
             TestRunner.RunTestScript("Test-VMUserDataBase64Encoded");
         }
 
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestVMNoPublicIPAddress()
+        {
+            TestRunner.RunTestScript("Test-VMNoPublicIPAddress");
+        }
+
     }
 }