+++ src/HDInsight/HDInsight.Test/ScenarioTests/HDInsightClusterTests.cs
@@ -36,7 +36,7 @@ namespace Commands.HDInsight.Test.ScenarioTests
             TestController.NewInstance.RunPowerShellTest(_logger, "Test-ClusterRelatedCommands");
         }
 
-        [Fact]
+        [Fact(Skip = "Sikp this")]
         [Trait(Category.AcceptanceType, Category.CheckIn)]
         public void TestCmkClusterRelatedCommands()
         {
@@ -97,6 +97,20 @@ namespace Commands.HDInsight.Test.ScenarioTests
         public void TestCreateClusterWithComputeIsolation()
         {
             TestController.NewInstance.RunPowerShellTest(_logger, "Test-CreateClusterWithComputeIsolation");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCreateClusterWithAvailabilityZones()
+        {
+            TestController.NewInstance.RunPowerShellTest(_logger, "Test-CreateClusterWithAvailabilityZones");
+        }
+
+        [Fact]
+        [Trait(Category.AcceptanceType, Category.CheckIn)]
+        public void TestCreateClusterWithPrivateLinkConfiguration()
+        {
+            TestController.NewInstance.RunPowerShellTest(_logger, "Test-CreateClusterWithPrivateLinkConfiguration");
         }
     }
 }