+++ src/Synapse/Synapse/Properties/AssemblyInfo.cs
@@ -49,5 +49,5 @@ using System.Runtime.InteropServices;
 // by using the '*' as shown below:
 
 
-[assembly: AssemblyVersion("0.18.0")]
-[assembly: AssemblyFileVersion("0.18.0")]
+[assembly: AssemblyVersion("0.19.0")]
+[assembly: AssemblyFileVersion("0.19.0")]