+++ src/Functions/ChangeLog.md
@@ -18,6 +18,12 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* [Breaking change] `Update-AzFunctionAppPlan` prompts for confirmation [#16490]
+* [Breaking change] `Remove-AzFunctionApp` does not delete ASP if it is the last app in the plan [#16487]
+* [Breaking change] Set the `FunctionsVersion` to 4 for FunctionApp creation [#16426]
+* [Breaking change] `Update-AzFunctionApp` prompts for confirmation [#14442]
+* Fixed an error creating function with `New-AzFunctionApp` on PowerShell 5.1 [#15430]
+* Supported storage account SKU `Standard_GZRS` [#14633]
 
 ## Version 3.1.1
 * Set site config netFrameworkVersion for Windows V4 apps only