+++ src/Functions/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+
+## Version 4.0.0
 * [Breaking change] `Update-AzFunctionAppPlan` prompts for confirmation [#16490]
 * [Breaking change] `Remove-AzFunctionApp` does not delete ASP if it is the last app in the plan [#16487]
 * [Breaking change] Set the `FunctionsVersion` to 4 for FunctionApp creation [#16426]